Verify passed sig in `sign_metadata` + refactor
By using python-tuf's verify_delegate instead of the custom signature
validation function used before, we only know if the threshold is
met or not and ignore any superfluous valid or invalid signatures.

Without additional check, we would add the signature passed to
`sign_metadata` even if it was invalid.

This patch adds an additional check, that only verifies the passed
signature (see _validate_signature), prior to overall threshold
verification.

In addition the patch refactors sign_metadata in an attempt to simplify
the control flow, in make future additions easier. E.g. adding support
for a "metadata update" signing event.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>